Update Profile.tsx
Import Link, User, UserAttribute, AuthService

Define type for ProfileState
Define type for ProfileProps

Declare component profile to have Props and State
Initialise State with userAttributes array
Asynchronously define that if component is mounted with UserAttributes from user
Then, setState with the new info.

Private define renderUserAttributes starting with rows as an empty array
Looping over each user attributes to create rows with key, name and value.
Return table with rows in the body
Define profileSpace
If there is a user, then show the content of the div, if not direct users to Login page
Return welcome to the profile page with the profileSpace component.